put window and file watching code in to its own module
All the window and file watching related code is now in its own module
so that multiple windows can be created. There are now IPC messages to
change the path of the markdown file that is rendered (i.e. render and
watch a different file in dthe same window), and to open a file in a new
window.